Def `useCurrentTime` hook; refactor
