+++ src/components/ControllerView/Clock.tsx
@@ -1,68 +1,13 @@
-import React, { HTMLAttributes, useEffect, useState } from 'react'
+import useCurrentTime from '$/hooks/useCurrentTime'
+import React, { HTMLAttributes } from 'react'
 import tw from 'twin.macro'
 
-type Time = {
-    ampm: 'am' | 'pm'
-    hour: string
-    minute: string
-}
-
-type Props = {
-    className?: string
-}
-
-export default function Clock({ className }: Props) {
-    const [time, setTime] = useState<Time>()
-
-    useEffect(() => {
-        let mounted = true
-
-        let recentTime: string
-
-        let timeout: number | undefined
-
-        function tick() {
-            const newTime = new Date().toLocaleString('en-US', {
-                hour: 'numeric',
-                minute: 'numeric',
-                hour12: true,
-            })
-
-            if (!mounted) {
-                return
-            }
-
-            if (recentTime !== newTime) {
-                recentTime = newTime
-
-                const [, hour = '', minute = '', ampm = 'am'] =
-                    newTime.toLowerCase().match(/^(\d+):(\d+) (am|pm)$/i) || []
-
-                setTime({
-                    ampm: ampm as Time['ampm'],
-                    hour,
-                    minute,
-                })
-            }
-
-            timeout = window.setTimeout(tick, 1000)
-        }
-
-        tick()
-
-        return () => {
-            if (timeout) {
-                clearTimeout(timeout)
-                timeout = undefined
-            }
-
-            mounted = false
-        }
-    }, [])
+export default function Clock(props: HTMLAttributes<HTMLDivElement>) {
+    const time = useCurrentTime()
 
     if (typeof time === 'undefined') {
         return (
-            <div className={className}>
+            <div {...props}>
                 {/* Preserve space. */}
                 &zwnj;
             </div>
@@ -71,6 +16,7 @@ export default function Clock({ className }: Props) {
 
     return (
         <div
+            {...props}
             css={[
                 tw`
                     select-none
@@ -81,10 +27,14 @@ export default function Clock({ className }: Props) {
                 `,
             ]}
         >
-            <div css={[tw`flex`]}>
-                <Digits>{time.hour}</Digits>
-                <Char css={[tw`w-[0.3em]`]}>:</Char>
-                <Digits>{time.minute}</Digits>
+            <div
+                css={[
+                    tw`
+                        flex
+                    `,
+                ]}
+            >
+                {time.hour}:{time.minute}
             </div>
             <div>
                 <span
@@ -98,52 +48,8 @@ export default function Clock({ className }: Props) {
                         `,
                     ]}
                 >
-                    {time.ampm}
+                    {time.period}
                 </span>
-            </div>
-        </div>
-    )
-}
-
-interface DigitsProps {
-    children: string
-}
-
-function Digits({ children }: DigitsProps) {
-    return (
-        <>
-            {children.split('').map((d, i) => (
-                <Char key={`${d}-${i}`}>{d}</Char>
-            ))}
-        </>
-    )
-}
-
-function Char({ children, ...props }: HTMLAttributes<HTMLDivElement>) {
-    return (
-        <div
-            {...props}
-            css={[
-                tw`
-                    w-[0.55em]
-                    text-center
-                    relative
-                `,
-            ]}
-        >
-            &zwnj;
-            <div
-                css={[
-                    tw`
-                        absolute
-                        top-1/2
-                        left-1/2
-                        -translate-x-1/2
-                        -translate-y-1/2
-                    `,
-                ]}
-            >
-                {children}
             </div>
         </div>
     )
